Extract device interactions from the trustchain SDK
